+++ src/Controller/TrainerController.php
@@ -5,10 +5,10 @@ namespace App\Controller;
 use App\Entity\Trainer;
 use App\Form\TrainerType;
 use Doctrine\ORM\EntityManagerInterface;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
 class TrainerController extends AbstractController
@@ -17,6 +17,9 @@ class TrainerController extends AbstractController
     #[IsGranted('IS_AUTHENTICATED_FULLY')]
     public function index(EntityManagerInterface $entityManager): Response
     {
+        if(!$this->getUser()->isVerified()){
+            return $this->redirectToRoute('app_home');
+        }
         //SELECT * FROM  trainer ORDER BY name ASC
         $trainers = $entityManager->getRepository(Trainer::class)->findBy([], ['name' => 'ASC']);
         return $this->render('trainer/index.html.twig', [
@@ -29,6 +32,9 @@ class TrainerController extends AbstractController
     #[IsGranted('IS_AUTHENTICATED_FULLY')]
     public function new(Trainer $trainer = null, Request $request, EntityManagerInterface $entityManager): Response
     {
+        if(!$this->getUser()->isVerified()){
+            return $this->redirectToRoute('app_home');
+        }
         if (!$this->isGranted("ROLE_ADMIN")) {
             return $this->redirectToRoute('app_trainer');
         } else {
@@ -62,6 +68,9 @@ class TrainerController extends AbstractController
     #[IsGranted('IS_AUTHENTICATED_FULLY')]
     public function remove(Trainer $trainer, EntityManagerInterface $entityManager): Response
     {
+        if(!$this->getUser()->isVerified()){
+            return $this->redirectToRoute('app_home');
+        }
         if ($this->isGranted("ROLE_ADMIN")) {
             $entityManager->remove($trainer);
             $entityManager->flush();
@@ -74,6 +83,9 @@ class TrainerController extends AbstractController
     #[IsGranted('IS_AUTHENTICATED_FULLY')]
     public function show(Trainer $trainer): Response
     {
+        if(!$this->getUser()->isVerified()){
+            return $this->redirectToRoute('app_home');
+        }
         return $this->render('trainer/show.html.twig', [
             'trainer' => $trainer
         ]);
